+++ compiler-tests/src/test/data/diagnostic/fir/provides/AssistedParametersAreNotSupportedOnProvides.kt
@@ -1,0 +1,6 @@
+// RENDER_DIAGNOSTICS_FULL_TEXT
+
+interface Example {
+  @Provides
+  fun provideString(<!PROVIDES_ERROR!>@Assisted<!> int: Int): String = "Hello, assisted parameters"
+}